+++ Models/ProductosModel.cs
@@ -4,5 +4,5 @@ public class Productos
 {
     public int IdProducto { get ; set; }
     public string Descripcion { get; set; }
-    public double Precio { get; set; }
+    public decimal Precio { get; set; }
 }
